Fix alignment of code blocks
These were supposed to be outdented, to match the preview blocks, but
was being overridden by the `margin` declaration just after.

I'm not convinced this is better (or worse), but now is working as
originally intended, and when we take a proper look at the design of
the component guide as a whole.