Add 'preview' feature, isolated view of component
Similar to the permalink pages for a single fixture, this feature
can show a component fixture without any application styles, or
page chrome (header/footer). This prevents any component-guide
specific styles affecting rendering.

This'll be useful for printing off examples of each component, wihout
having to manually crop out wrapping UI.

It can also show a list of all the fixtures for a single component, which
is useful for comparing all the examples side-by-side easily and quickly.