Add a startup script, with fixed port
This app isn't part of our production stack, but it would be useful to give
it a fixed port, rather than having to run on 3000 and risk conflicting
with other rails apps.

I've had this with govuk-prototyping-rails, which is common to run along
side the component guide.

There is an equivilent PR on the private development repo [1] to
reserve the port used here.

[1] https://github.gds/gds/development/pull/262