Rename `.preview` to avoid clashes with component
Adding attachment HTML to the govspeak component includes examples
of `preview` as a class name, which was being overridden by the
component guide styling.

This has come up a few times with the component guide, and we need
a better long term fix, but for now this fixes the attachment examples
i'm working on.

Longer term fixes might include;
 - using a BEM like approach in components, to avoid class name clashes
 - using BEM in the component guide and avoiding element level selectors
   entirely.

https://github.com/alphagov/govuk-component-guide/pull/16