Return 404 error for unknown components
At the moment, there is no error checking when requesting components from the component guide. Therefore, if an unknown component is requested, the user sees the default Rails 500 error page rather than the more appropriate, branded 404 error.

This fix carries out error checking when requesting a component, and returns a 404 error page when the requested component is not found.